Release/2.2.2.1

+++ django-os2webscanner/os2webscanner/models/statistic_model.py
@@ -49,4 +49,4 @@ class TypeStatistics(models.Model):
 
     type_name = models.CharField(max_length=256)
     count = models.IntegerField(default=0)
-    size = models.IntegerField(default=0)
+    size = models.BigIntegerField(default=0)
